requirements.txt
python-dotenv
anthropic
modal